Base logger cleanup
